+++ lib/core/providers/analytics_provider.dart
@@ -1,0 +1,8 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+import '../services/analytics_service.dart';
+import '../services/shared_prefs_service.dart';
+
+final analyticsServiceProvider = Provider<AnalyticsService>((ref) {
+  return AnalyticsService(ref.watch(sharedPrefsServiceProvider));
+});